Moved integration tests extensions into one folder.
Renamed CoreNodeExtensions to FullNodeTestBuilderExtension.  Looks like it was committed with the wrong name.

Used extension CoreNode method GetProofOfWorkRewardForMinedBlocks() to calculate the reward by taking SubsidyHalvingInterval into account.
